Function and file renames accordingly
